fix estrutura
